Added PathBreadcrumbViewModel
