+++ src/Files.App/UserControls/InnerNavigationToolbar.xaml.cs
@@ -30,15 +30,15 @@ namespace Files.App.UserControls
 
 		public readonly PreviewPaneViewModel PreviewPaneViewModel;
 
-		public ToolbarViewModel ViewModel
+		public AddressToolbarViewModel ViewModel
 		{
-			get => (ToolbarViewModel)GetValue(ViewModelProperty);
+			get => (AddressToolbarViewModel)GetValue(ViewModelProperty);
 			set => SetValue(ViewModelProperty, value);
 		}
 
 		// Using a DependencyProperty as the backing store for ViewModel.  This enables animation, styling, binding, etc...
 		public static readonly DependencyProperty ViewModelProperty =
-			DependencyProperty.Register(nameof(ViewModel), typeof(ToolbarViewModel), typeof(InnerNavigationToolbar), new PropertyMetadata(null));
+			DependencyProperty.Register(nameof(ViewModel), typeof(AddressToolbarViewModel), typeof(InnerNavigationToolbar), new PropertyMetadata(null));
 
 		public bool ShowViewControlButton
 		{
