+++ src/Files.App/ViewModels/UserControls/AddressToolbarViewModel.cs
@@ -10,43 +10,28 @@ using Microsoft.UI.Xaml.Controls.Primitives;
 using Microsoft.UI.Xaml.Input;
 using System.IO;
 using System.Windows.Input;
-using Windows.ApplicationModel.DataTransfer;
-using Windows.UI.Text;
 
 namespace Files.App.ViewModels.UserControls
 {
 	/// <summary>
 	/// Represents ViewModel for <see cref="Files.App.UserControls.AddressToolbar"/>.
 	/// </summary>
-	public class ToolbarViewModel : ObservableObject, IAddressToolbar, IDisposable
+	public class AddressToolbarViewModel : ObservableObject, IAddressToolbar, IDisposable
 	{
 		// Dependency injection
 
 		private IUserSettingsService UserSettingsService { get; } = Ioc.Default.GetRequiredService<IUserSettingsService>();
-		private MainPageViewModel MainPageViewModel { get; } = Ioc.Default.GetRequiredService<MainPageViewModel>();
+		private IContentPageContext ContentPageContext { get; } = Ioc.Default.GetRequiredService<IContentPageContext>();
 		private DrivesViewModel DrivesViewModel { get; } = Ioc.Default.GetRequiredService<DrivesViewModel>();
 		private IDialogService DialogService { get; } = Ioc.Default.GetRequiredService<IDialogService>();
-		private IUpdateService UpdateService { get; } = Ioc.Default.GetService<IUpdateService>()!;
+		private IUpdateService UpdateService { get; } = Ioc.Default.GetRequiredService<IUpdateService>();
 		private ICommandManager Commands { get; } = Ioc.Default.GetRequiredService<ICommandManager>();
-
-		// Event delegates
-
-		public delegate void ToolbarPathItemInvokedEventHandler(object sender, PathNavigationEventArgs e);
-		public delegate void ToolbarPathItemLoadedEventHandler(object sender, ToolbarPathItemLoadedEventArgs e);
-		public delegate void AddressBarTextEnteredEventHandler(object sender, AddressBarTextEnteredEventArgs e);
-		public delegate void ToolbarFlyoutOpenedEventHandler(object sender, ToolbarFlyoutOpenedEventArgs e);
-		public delegate void PathBoxItemDroppedEventHandler(object sender, PathBoxItemDroppedEventArgs e);
 
 		// Event handlers
 
-		public event IAddressToolbar.ItemDraggedOverPathItemEventHandler? ItemDraggedOverPathItem;
+		public delegate void AddressBarTextEnteredEventHandler(object sender, AddressBarTextEnteredEventArgs e);
 		public event IAddressToolbar.ToolbarQuerySubmittedEventHandler? PathBoxQuerySubmitted;
-		public event ToolbarPathItemInvokedEventHandler? ToolbarPathItemInvoked;
-		public event ToolbarFlyoutOpenedEventHandler? ToolbarFlyoutOpened;
-		public event ToolbarPathItemLoadedEventHandler? ToolbarPathItemLoaded;
 		public event AddressBarTextEnteredEventHandler? AddressBarTextEntered;
-		public event PathBoxItemDroppedEventHandler? PathBoxItemDropped;
-		private PointerRoutedEventArgs? pointerRoutedEventArgs;
 		public event EventHandler? EditModeEnabled;
 		public event EventHandler? RefreshRequested;
 		public event EventHandler? RefreshWidgetsRequested;
@@ -155,7 +140,7 @@ namespace Files.App.ViewModels.UserControls
 			set => SetProperty(ref _PathText, value);
 		}
 
-		private CurrentInstanceViewModel _InstanceViewModel;
+		private CurrentInstanceViewModel? _InstanceViewModel;
 		public CurrentInstanceViewModel InstanceViewModel
 		{
 			get => _InstanceViewModel;
@@ -172,7 +157,7 @@ namespace Files.App.ViewModels.UserControls
 			}
 		}
 
-		private Style _LayoutOpacityIcon;
+		private Style? _LayoutOpacityIcon;
 		public Style LayoutOpacityIcon
 		{
 			get => _LayoutOpacityIcon;
@@ -231,7 +216,7 @@ namespace Files.App.ViewModels.UserControls
 			set => SetProperty(ref _ClickablePathLoaded, value);
 		}
 
-		private string _PathControlDisplayText;
+		private string? _PathControlDisplayText;
 		public string PathControlDisplayText
 		{
 			get => _PathControlDisplayText;
@@ -240,9 +225,7 @@ namespace Files.App.ViewModels.UserControls
 
 		// Auto properties
 
-		public ObservableCollection<PathBoxItem> PathComponents { get; } = new();
-
-		public ObservableCollection<NavigationBarSuggestionItem> NavigationBarSuggestions = new();
+		public ObservableCollection<NavigationBarSuggestionItem> NavigationBarSuggestions { get; } = new();
 
 		public bool IsEditModeEnabled
 		{
@@ -271,14 +254,13 @@ namespace Files.App.ViewModels.UserControls
 		public SearchBoxViewModel SearchBoxViewModel
 			=> (SearchBoxViewModel)SearchBox;
 
-		public ICommand? OpenNewWindowCommand { get; set; }
-		public ICommand? CreateNewFileCommand { get; set; }
-		public ICommand? Share { get; set; }
-		public ICommand? UpdateCommand { get; set; }
+		public ICommand? OpenNewWindowCommand { get; }
+		public ICommand? CreateNewFileCommand { get; }
+		public ICommand? Share { get; }
+		public ICommand? UpdateCommand { get; }
 		public ICommand? RefreshClickCommand { get; }
 		public ICommand? ViewReleaseNotesAsyncCommand { get; }
 
-		public bool IsSingleItemOverride { get; set; } = false;
 		public bool SearchHasFocus { get; private set; }
 
 		public bool HasAdditionalAction => InstanceViewModel.IsPageTypeRecycleBin || IsPowerShellScript || CanExtract || IsImage || IsFont || IsInfFile;
@@ -310,20 +292,15 @@ namespace Files.App.ViewModels.UserControls
 		private static AddressToolbar? AddressToolbar
 			=> (MainWindow.Instance.Content as Frame)?.FindDescendant<AddressToolbar>();
 
-		private readonly DispatcherQueue _dispatcherQueue;
-		private readonly DispatcherQueueTimer _dragOverTimer;
-		private string? _dragOverPath = null;
-		private bool _lockFlag = false;
-
 		// Methods
 
-		public ToolbarViewModel()
+		public AddressToolbarViewModel()
 		{
-			RefreshClickCommand = new RelayCommand<RoutedEventArgs>(e => RefreshRequested?.Invoke(this, EventArgs.Empty));
 			ViewReleaseNotesAsyncCommand = new AsyncRelayCommand(ViewReleaseNotesAsync);
-
-			_dispatcherQueue = DispatcherQueue.GetForCurrentThread();
-			_dragOverTimer = _dispatcherQueue.CreateTimer();
+			CreateNewFileCommand = new AsyncRelayCommand<ShellNewEntry>(x => UIFilesystemHelpers.CreateFileFromDialogResultTypeAsync(AddItemDialogItemType.File, x, ContentPageContext.ShellPage));
+			OpenNewWindowCommand = new AsyncRelayCommand(NavigationHelpers.LaunchNewWindowAsync);
+			RefreshClickCommand = new RelayCommand<RoutedEventArgs>(e => RefreshRequested?.Invoke(this, EventArgs.Empty));
+			UpdateCommand = new AsyncRelayCommand(UpdateService.DownloadUpdatesAsync);
 
 			SearchBox.Escaped += SearchRegion_Escaped;
 			UserSettingsService.OnSettingChangedEvent += UserSettingsService_OnSettingChangedEvent;
@@ -351,7 +328,7 @@ namespace Files.App.ViewModels.UserControls
 			await dialog.TryShowAsync();
 		}
 
-		public async Task CheckForReleaseNotesAsync()
+		private async Task CheckForReleaseNotesAsync()
 		{
 			var result = await UpdateService.GetLatestReleaseNotesAsync();
 			if (result is null)
@@ -359,11 +336,6 @@ namespace Files.App.ViewModels.UserControls
 
 			ReleaseNotes = result;
 			IsReleaseNotesVisible = true;
-		}
-
-		public void RefreshWidgets()
-		{
-			RefreshWidgetsRequested?.Invoke(this, EventArgs.Empty);
 		}
 
 		private void UserSettingsService_OnSettingChangedEvent(object? sender, SettingChangedEventArgs e)
@@ -381,146 +353,6 @@ namespace Files.App.ViewModels.UserControls
 			}
 		}
 
-		public void PathBoxItem_DragLeave(object sender, DragEventArgs e)
-		{
-			if (((StackPanel)sender).DataContext is not PathBoxItem pathBoxItem ||
-				pathBoxItem.Path == "Home")
-			{
-				return;
-			}
-
-			// Reset dragged over pathbox item
-			if (pathBoxItem.Path == _dragOverPath)
-				_dragOverPath = null;
-		}
-
-		public async Task PathBoxItem_Drop(object sender, DragEventArgs e)
-		{
-			if (_lockFlag)
-				return;
-
-			_lockFlag = true;
-
-			// Reset dragged over pathbox item
-			_dragOverPath = null;
-
-			if (((StackPanel)sender).DataContext is not PathBoxItem pathBoxItem ||
-				pathBoxItem.Path == "Home")
-			{
-				return;
-			}
-
-			var deferral = e.GetDeferral();
-
-			var signal = new AsyncManualResetEvent();
-
-			PathBoxItemDropped?.Invoke(this, new PathBoxItemDroppedEventArgs()
-			{
-				AcceptedOperation = e.AcceptedOperation,
-				Package = e.DataView,
-				Path = pathBoxItem.Path,
-				SignalEvent = signal
-			});
-
-			await signal.WaitAsync();
-
-			deferral.Complete();
-			await Task.Yield();
-
-			_lockFlag = false;
-		}
-
-		public async Task PathBoxItem_DragOver(object sender, DragEventArgs e)
-		{
-			if (IsSingleItemOverride ||
-				((StackPanel)sender).DataContext is not PathBoxItem pathBoxItem ||
-				pathBoxItem.Path == "Home")
-			{
-				return;
-			}
-
-			if (_dragOverPath != pathBoxItem.Path)
-			{
-				_dragOverPath = pathBoxItem.Path;
-				_dragOverTimer.Stop();
-
-				if (_dragOverPath != (this as IAddressToolbar).PathComponents.LastOrDefault()?.Path)
-				{
-					_dragOverTimer.Debounce(() =>
-					{
-						if (_dragOverPath is not null)
-						{
-							_dragOverTimer.Stop();
-							ItemDraggedOverPathItem?.Invoke(this, new PathNavigationEventArgs()
-							{
-								ItemPath = _dragOverPath
-							});
-							_dragOverPath = null;
-						}
-					},
-					TimeSpan.FromMilliseconds(1000), false);
-				}
-			}
-
-			// In search page
-			if (!FilesystemHelpers.HasDraggedStorageItems(e.DataView) || string.IsNullOrEmpty(pathBoxItem.Path))
-			{
-				e.AcceptedOperation = DataPackageOperation.None;
-
-				return;
-			}
-
-			e.Handled = true;
-			var deferral = e.GetDeferral();
-
-			var storageItems = await FilesystemHelpers.GetDraggedStorageItems(e.DataView);
-
-			if (!storageItems.Any(storageItem =>
-					!string.IsNullOrEmpty(storageItem?.Path) &&
-					storageItem.Path.Replace(pathBoxItem.Path, string.Empty, StringComparison.Ordinal)
-						.Trim(Path.DirectorySeparatorChar)
-						.Contains(Path.DirectorySeparatorChar)))
-			{
-				e.AcceptedOperation = DataPackageOperation.None;
-			}
-
-			// Copy be default when dragging from zip
-			else if (storageItems.Any(x =>
-					x.Item is ZipStorageFile ||
-					x.Item is ZipStorageFolder) ||
-					ZipStorageFolder.IsZipPath(pathBoxItem.Path))
-			{
-				e.DragUIOverride.Caption = string.Format("CopyToFolderCaptionText".GetLocalizedResource(), pathBoxItem.Title);
-				e.AcceptedOperation = DataPackageOperation.Copy;
-			}
-			else
-			{
-				e.DragUIOverride.IsCaptionVisible = true;
-				e.DragUIOverride.Caption = string.Format("MoveToFolderCaptionText".GetLocalizedResource(), pathBoxItem.Title);
-				e.AcceptedOperation = DataPackageOperation.Move;
-			}
-
-			deferral.Complete();
-		}
-
-		public void PathItemSeparator_DataContextChanged(FrameworkElement sender, DataContextChangedEventArgs args)
-		{
-			var pathSeparatorIcon = sender as FontIcon;
-			if (pathSeparatorIcon is null || pathSeparatorIcon.DataContext is null)
-				return;
-
-			ToolbarPathItemLoaded?.Invoke(pathSeparatorIcon, new ToolbarPathItemLoadedEventArgs()
-			{
-				Item = (PathBoxItem)pathSeparatorIcon.DataContext,
-				OpenedFlyout = (MenuFlyout)pathSeparatorIcon.ContextFlyout
-			});
-		}
-
-		public void PathBoxItemFlyout_Opened(object sender, object e)
-		{
-			ToolbarFlyoutOpened?.Invoke(this, new ToolbarFlyoutOpenedEventArgs() { OpenedFlyout = (MenuFlyout)sender });
-		}
-
 		public void VisiblePath_TextChanged(AutoSuggestBox sender, AutoSuggestBoxTextChangedEventArgs args)
 		{
 			if (args.Reason == AutoSuggestionBoxTextChangeReason.UserInput)
@@ -532,39 +364,6 @@ namespace Files.App.ViewModels.UserControls
 			PathBoxQuerySubmitted?.Invoke(this, new ToolbarQuerySubmittedEventArgs() { QueryText = args.QueryText });
 
 			(this as IAddressToolbar).IsEditModeEnabled = false;
-		}
-
-		public void PathBoxItem_PointerPressed(object sender, PointerRoutedEventArgs e)
-		{
-			if (e.Pointer.PointerDeviceType != Microsoft.UI.Input.PointerDeviceType.Mouse)
-				return;
-
-			var ptrPt = e.GetCurrentPoint(AddressToolbar);
-			pointerRoutedEventArgs = ptrPt.Properties.IsMiddleButtonPressed ? e : null;
-		}
-
-		public async Task PathBoxItem_Tapped(object sender, TappedRoutedEventArgs e)
-		{
-			var itemTappedPath = ((sender as TextBlock)?.DataContext as PathBoxItem)?.Path;
-			if (itemTappedPath is null)
-				return;
-
-			if (pointerRoutedEventArgs is not null)
-			{
-				await MainWindow.Instance.DispatcherQueue.EnqueueOrInvokeAsync(async () =>
-				{
-					await MainPageViewModel.AddNewTabByPathAsync(typeof(PaneHolderPage), itemTappedPath);
-				}, DispatcherQueuePriority.Low);
-				e.Handled = true;
-				pointerRoutedEventArgs = null;
-
-				return;
-			}
-
-			ToolbarPathItemInvoked?.Invoke(this, new PathNavigationEventArgs()
-			{
-				ItemPath = itemTappedPath
-			});
 		}
 
 		public void OpenCommandPalette()
@@ -641,67 +440,6 @@ namespace Files.App.ViewModels.UserControls
 
 		private void SearchRegion_Escaped(object? sender, ISearchBox searchBox)
 			=> CloseSearchBox(true);
-
-		public async Task SetPathBoxDropDownFlyoutAsync(MenuFlyout flyout, PathBoxItem pathItem, IShellPage shellPage)
-		{
-			var nextPathItemTitle = PathComponents[PathComponents.IndexOf(pathItem) + 1].Title;
-			IList<StorageFolderWithPath>? childFolders = null;
-
-			StorageFolderWithPath folder = await shellPage.FilesystemViewModel.GetFolderWithPathFromPathAsync(pathItem.Path);
-			if (folder is not null)
-				childFolders = (await FilesystemTasks.Wrap(() => folder.GetFoldersWithPathAsync(string.Empty))).Result;
-
-			flyout.Items?.Clear();
-
-			if (childFolders is null || childFolders.Count == 0)
-			{
-				var flyoutItem = new MenuFlyoutItem
-				{
-					Icon = new FontIcon { Glyph = "\uE7BA" },
-					Text = "SubDirectoryAccessDenied".GetLocalizedResource(),
-					//Foreground = (SolidColorBrush)Application.Current.Resources["SystemControlErrorTextForegroundBrush"],
-					FontSize = 12
-				};
-
-				flyout.Items?.Add(flyoutItem);
-
-				return;
-			}
-
-			var boldFontWeight = new FontWeight { Weight = 800 };
-			var normalFontWeight = new FontWeight { Weight = 400 };
-
-			var workingPath =
-				PathComponents[PathComponents.Count - 1].Path?.TrimEnd(Path.DirectorySeparatorChar);
-
-			foreach (var childFolder in childFolders)
-			{
-				var isPathItemFocused = childFolder.Item.Name == nextPathItemTitle;
-
-				var flyoutItem = new MenuFlyoutItem
-				{
-					Icon = new FontIcon
-					{
-						Glyph = "\uED25",
-						FontWeight = isPathItemFocused ? boldFontWeight : normalFontWeight
-					},
-					Text = childFolder.Item.Name,
-					FontSize = 12,
-					FontWeight = isPathItemFocused ? boldFontWeight : normalFontWeight
-				};
-
-				if (workingPath != childFolder.Path)
-				{
-					flyoutItem.Click += (sender, args) =>
-					{
-						// Navigate to the directory
-						shellPage.NavigateToPath(childFolder.Path);
-					};
-				}
-
-				flyout.Items?.Add(flyoutItem);
-			}
-		}
 
 		public async Task CheckPathInputAsync(string currentInput, string currentSelectedPath, IShellPage shellPage)
 		{
