+++ src/Files.App/ViewModels/UserControls/IAddressToolbarViewModel.cs
@@ -5,7 +5,7 @@ using Files.App.Views;
 using Microsoft.UI.Xaml.Controls;
 using Windows.ApplicationModel.DataTransfer;
 
-namespace Files.App.UserControls
+namespace Files.App.ViewModels.UserControls
 {
 	public interface IAddressToolbar
 	{
@@ -28,11 +28,7 @@ namespace Files.App.UserControls
 
 		public bool CanGoForward { get; set; }
 
-		public bool IsSingleItemOverride { get; set; }
-
 		public string PathControlDisplayText { get; set; }
-
-		public ObservableCollection<PathBoxItem> PathComponents { get; }
 
 		public delegate void ToolbarQuerySubmittedEventHandler(object sender, ToolbarQuerySubmittedEventArgs e);
 
@@ -41,8 +37,6 @@ namespace Files.App.UserControls
 		public event ToolbarQuerySubmittedEventHandler PathBoxQuerySubmitted;
 
 		public event EventHandler EditModeEnabled;
-
-		public event ItemDraggedOverPathItemEventHandler ItemDraggedOverPathItem;
 
 		public event EventHandler RefreshRequested;
